pyproject.toml
[tool.poetry]
name = "langchain-basics"
version = "0.1.0"
description = ""
authors = ["Caio Ramos <cwfr88@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.11.4"
langchain = "^0.2.11"
langchain-openai = "^0.1.17"
python-dotenv = "^1.0.1"
langchain-community = "^0.2.10"
pypdf = "^4.3.1"
wikipedia = "^1.4.0"
langchain-chroma = "^0.1.2"
faiss-cpu = "^1.8.0.post1"
bs4 = "^0.0.2"
unstructured = "^0.15.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
